WIP: adding SQLite backend
